add location feature

+++ database/migrations/2022_10_20_123006_create_restaurants_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('image');
             $table->string('location')->nullable();
-            $table->decimal('latitude', 11, 8);
-            $table->decimal('longitude', 11, 8);
+            $table->float('latitude', 10, 6);
+            $table->float('longitude', 10, 6);
             $table->decimal('delivery_fee');
             $table->enum('status', ['Available', 'Closed','Busy'])->default('Available');
             $table->foreignId('category_id')->constrained('categories')->onUpdate('cascade')
